Renamed OfficeSpace.py to RectangleOverlap.py
I will probably not be using the code that I have written in my submission. The program is a lot easier of we know that the rectangle coordinates are integers (we can just cycle through the x and y values in a for loop... much easier).

With that siad, I Want to continue developing my RectangleOverlap file, since I think that it's a very interesting problem. Expect a lot of progress tomorrow.