Added a rectangle + fixed length bug
I added a rectangle class. This class has a few methods to check if the rectangle exists to the right of a given x value. This will be useful in finding the area.

I also added some tests for the rectangle class.

I also found a bug in my earlier code. the len function must return an integer. As such, my __len__ methods for the interval and interval_union classes would only work if the length came out to an integer. oops. I got rid of this method and instead defined a length method.

Realizing this made me redo many of the tests to use floats (and unsymmetric ones at that) to improve the testing of my code.

As of right now, all the tests work (yay!). I next plan on writing the area calculating code (and then writing the code for this homework).