Added in a "read in" method + fixed a few bugs
I added a method to read in cases from file. This code appears to work. I also fixed a few bugs in the printing parts of the Rectangle and Case class.

Unfortuneatly... the code is not finished. What I print out for the employee areas is wrong. I need to print out the "uncontested area" (basically, the area that the employee claimed that is not being claimed by another employee). I need to write a function to figure that out before I can call the code "done"