Added method to calculate guaranteed area
This new method calculates the uncontested area for each employee. This is used when printing the Case. The code is pretty similar to the code in calculate_contested_area... I think I will write a "sweep" method (that both call) to do this instead. Expect that in a future commit.