Wrote a area calculation method (to eliminate identical code)
I wrote a new method, _calculate_requested_area. This method finds the area covered by the employees request in a given region.
This method was written to replace the nearly identical code that was in _calculate_guaranteed_area and _calculate_allocated_area. The new method appears to work (the values all look correct!) and cut down the overall length of the program (more yay!).